% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-V1.R
\name{fwa_bbox}
\alias{fwa_bbox}
\title{BBox}
\usage{
fwa_bbox(
  table,
  filter = NULL,
  geom_column = "geom",
  epsg = 4326,
  schema = "whse_basemapping"
)
}
\arguments{
\item{filter}{A named vector or list of the filter(s) to apply, where the list
names correspond to column names and the list values correspond to the desired
value, e.g. \code{list(gnis_name = "Sangan River")}.}

\item{epsg}{A positive whole number of the epsg to transform features to.}

\item{schema}{A character string of the schema to prepend to collection_id.
Possible values include 'whse_basemapping', 'usgs', 'hydrosheds' and NULL. If NULL, nothing is prepended.}
}
\value{
A sf object
}
\description{
Get bounding box of feature(s)
}
\examples{
fwa_bbox("fwa_stream_networks_sp", filter = "gnis_name = 'Cowichan River'")
}
