% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{fwa_collection}
\alias{fwa_collection}
\title{FWA collection}
\usage{
fwa_collection(
  collection_id,
  filter = NULL,
  limit = 100,
  offset = 0,
  bbox = NULL,
  properties = NULL,
  transform = NULL,
  epsg = 4326
)
}
\arguments{
\item{collection_id}{A character string of the collection id.}

\item{filter}{A named vector or list of the filter(s) to apply, where the list
names correspond to column names and the list values correspond to the desired
value, e.g. \code{list(gnis_name = "Sangan River")}.}

\item{limit}{A positive whole number indicating the maximum number of features to return.}

\item{offset}{A positive whole number indicating the offset of start of returned results.}

\item{bbox}{A vector of four numbers indicating bounding box to limit output
features to, e.g. \code{c(minLon, minLat, maxLon, maxLat)}.}

\item{properties}{A vector of strings of the column names to include. If NULL (default), all columns are retained.}

\item{transform}{A character vector with the name of the transform function followed by the parameter values (e.g. c("ST_Simplify", 100))}

\item{epsg}{A positive whole number of the epsg to transform features to.}
}
\value{
An sf object
}
\description{
Get and filter collection from database.
}
\examples{
\dontrun{
collection_id <- "whse_basemapping.fwa_stream_networks_sp"
fwa_collection(collection_id, filter = list(gnis_name = "Sangan River"))
}
}
\seealso{
\href{https://www.hillcrestgeo.ca/fwapg/collections.html}{collections API documentation}.
}
