% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection-properties.R, R/meta-properties.R
\name{fwa_collection_properties}
\alias{fwa_collection_properties}
\alias{fwa_meta_properties}
\title{Freshwater Atlas Collection's Properties}
\usage{
fwa_collection_properties(collection_id)

fwa_meta_properties(collection_id)
}
\arguments{
\item{collection_id}{A character string of the collection id.}
}
\value{
A tibble.
}
\description{
Get information about a collection's column names.
}
\section{Functions}{
\itemize{
\item \code{fwa_meta_properties()}: Freshwater Atlas Collection's Properties

\lifecycle{soft-deprecated}

}}
\examples{
fwa_collection_properties("fwapg.fwa_stream_networks_sp")
}
\seealso{
Other collections: 
\code{\link{fwa_collections}()},
\code{\link{fwa_collection}()}
}
\concept{collections}
