% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{fwa_collections}
\alias{fwa_collections}
\title{FWA Collections}
\usage{
fwa_collections()
}
\value{
A tibble.
}
\description{
Get information about the collections.
}
\examples{
fwa_collections()
}
\seealso{
Other metadata: 
\code{\link{fwa_collections_properties}()},
\code{\link{fwa_meta_collections}()},
\code{\link{fwa_meta_properties}()}
}
\concept{metadata}
