% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection-properties.R
\name{fwa_collections_properties}
\alias{fwa_collections_properties}
\title{FWA Collection's Properties}
\usage{
fwa_collections_properties(collection_id)
}
\arguments{
\item{collection_id}{A character string of the collection id.}
}
\value{
A tibble.
}
\description{
Get information about a collection's column names.
}
\examples{
fwa_collections_properties("whse_basemapping.fwa_stream_networks_sp")
}
\seealso{
Other metadata: 
\code{\link{fwa_collections}()},
\code{\link{fwa_meta_collections}()},
\code{\link{fwa_meta_properties}()}
}
\concept{metadata}
