% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{fwa_feature}
\alias{fwa_feature}
\title{FWA feature}
\usage{
fwa_feature(
  table,
  filter = NULL,
  columns = NULL,
  bounds = NULL,
  geom_column = "geom",
  epsg = 4326,
  schema = "whse_basemapping"
)
}
\arguments{
\item{table}{A character string of the table name.}

\item{filter}{A character string of the SQL WHERE statement to filter features by.}

\item{columns}{A vector of strings of the column names to include.}

\item{bounds}{A vector of numbers indicating bounding box to limit output features to, e.g. \code{c(sw.lng, sw.lat, ne.lng, ne.lat)}.}

\item{geom_column}{A character string of the name of the geometry column of the table.}

\item{epsg}{An integer of the epsg to transform features to.}

\item{schema}{A character string of the schema. Default 'whse_basemapping' is used for most layers.}
}
\value{
A sf object
}
\description{
Get features from table.
}
\examples{
fwa_feature("fwa_stream_networks_sp", filter = "gnis_name = 'Sangan River'")
}
