% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{fwa_function}
\alias{fwa_function}
\title{FWA Function}
\usage{
fwa_function(
  function_id,
  parameters,
  limit = 100,
  offset = 0,
  bbox = NULL,
  properties = NULL,
  transform = NULL,
  epsg = 4326
)
}
\arguments{
\item{function_id}{A character string of the function id.}

\item{parameters}{A named list of the parameters and values for a specific function.}

\item{limit}{A positive whole number indicating the maximum number of features to return.}

\item{offset}{A positive whole number indicating the offset of start of returned results.}

\item{bbox}{A vector of four numbers indicating bounding box to limit output
features to, e.g. \code{c(minLon, minLat, maxLon, maxLat)}.}

\item{properties}{A vector of strings of the column names to include. If NULL (default), all columns are retained.}

\item{transform}{A character vector with the name of the transform function followed by the parameter values (e.g. c("ST_Simplify", 100))}

\item{epsg}{A positive whole number of the epsg to transform features to.}
}
\value{
An sf object.
}
\description{
Execute general spatial function.
}
\seealso{
\href{https://www.hillcrestgeo.ca/fwapg/functions.html}{API documentation}
}
