% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{fwa_gnis_stream_network}
\alias{fwa_gnis_stream_network}
\title{FWA feature}
\usage{
fwa_gnis_stream_network(
  gnis_name,
  columns = NULL,
  bounds = NULL,
  geom_column = "geom",
  epsg = 4326
)
}
\arguments{
\item{gnis_name}{A character string of the gnis_name.}

\item{columns}{A vector of strings of the column names to include.}

\item{bounds}{A vector of numbers indicating bounding box to limit output features to, e.g. \code{c(sw.lng, sw.lat, ne.lng, ne.lat)}.}

\item{geom_column}{A character string of the name of the geometry column of the table.}

\item{epsg}{An integer of the epsg to transform features to.}
}
\value{
A sf object
}
\description{
Get features from table.
}
\examples{
fwa_gnis_stream_network('Sangan River')
}
