% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{fwa_gnis_streams}
\alias{fwa_gnis_streams}
\title{FWA stream network}
\usage{
fwa_gnis_streams(
  gnis_name,
  columns = NULL,
  bounds = NULL,
  geom_column = "geom",
  epsg = 4326,
  named_streams = FALSE
)
}
\arguments{
\item{gnis_name}{A character string of the gnis_name.}

\item{columns}{A vector of strings of the column names to include.}

\item{epsg}{A positive whole number of the epsg to transform features to.}

\item{named_streams}{A flag indicating whether to get data from the 'fwa_named_streams' layer.
If FALSE, the 'fwa_stream_networks_sp' layer is used.}
}
\value{
A sf object
}
\description{
Get stream network features from gnis_name.
}
\examples{
fwa_gnis_streams('Sangan River')
}
