% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fwa_index_point}
\alias{fwa_index_point}
\title{FWA index point}
\usage{
fwa_index_point(x, y, srid, tolerance = 5000, num_features = 1, epsg = srid)
}
\arguments{
\item{x}{A number of the x coordinate.}

\item{y}{A number of the y coordinate.}

\item{srid}{An integer of the srid of the coordinate (e.g. epsg).}

\item{tolerance}{A number of the tolerance.}

\item{num_features}{An integer of the maximum number of features to return.}

\item{epsg}{An integer of the epsg to transform results to.}
}
\value{
A sf object
}
\description{
Provided a point (x, y, srid), return the point along a stream(s) nearest to the point provided within specified tolerance.
}
\examples{
fwa_index_point(x = -132.26, y = 53.36, srid = 4326)
}
