% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-meta.R
\name{fwa_list_columns}
\alias{fwa_list_columns}
\title{FWA list columns}
\usage{
fwa_list_columns(table)
}
\arguments{
\item{table}{A character string of the table name.}
}
\value{
A tibble
}
\description{
List columns in a table.
}
\examples{
fwa_list_columns("fwa_stream_networks_sp")
}
