% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{fwa_list_layers}
\alias{fwa_list_layers}
\title{FWA list layers}
\usage{
fwa_list_layers()
}
\value{
A tibble
}
\description{
List available layers. \lifecycle{defunct}
}
\examples{
\dontrun{
fwa_list_layers()
}
}
