% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{fwa_list_tables}
\alias{fwa_list_tables}
\title{FWA list tables}
\usage{
fwa_list_tables()
}
\value{
A tibble
}
\description{
List available tables in data collections.
}
\examples{
\dontrun{fwa_list_tables()}
}
