% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fwa_locate_along_interval}
\alias{fwa_locate_along_interval}
\title{FWA locate along interval}
\usage{
fwa_locate_along_interval(
  blue_line_key,
  interval_length = 1000,
  start = 0,
  epsg = 4326
)
}
\arguments{
\item{blue_line_key}{An integer of the stream blue line key.}

\item{interval_length}{An integer of the interval distance in meters.}

\item{start}{An integer of the distance in meters upstream from the river mouth to start from.}

\item{epsg}{A positive whole number of the epsg to transform features to.}

\item{limit}{An integer of the limit of rows to return.}
}
\value{
A sf object
}
\description{
Provided a blue_line_key and starting downstream distance, return a set of points along an interval.
}
\examples{
fwa_locate_along_interval(356308001, interval_length = 10, start = 0)
}
