% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{fwa_meta_collections}
\alias{fwa_meta_collections}
\title{FWA collections metadata}
\usage{
fwa_meta_collections()
}
\value{
A tibble
}
\description{
Provide collection metadata.
}
\examples{
\dontrun{fwa_meta_collections()}
}
