% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwa_network_trace.R
\name{fwa_network_trace}
\alias{fwa_network_trace}
\title{Freshwater Atlas Network Trace}
\usage{
fwa_network_trace(
  blue_line_key_a,
  measure_a = 0,
  blue_line_key_b,
  measure_b = 0,
  tolerance = 1,
  aggregate_path = TRUE,
  epsg = 4326
)
}
\arguments{
\item{tolerance}{A number of the tolerance in m.}

\item{epsg}{A positive whole number of the epsg to transform features to.}
}
\value{
An sf object.
}
\description{
Returns the stream network path between the provided locations
}
\examples{
fwa_network_trace(356308001, downstream_route_measure = 10000)
}
\seealso{
\href{https://features.hillcrestgeo.ca/fwa/functions/fwa_networktrace.html}{API documentation}

Other functions: 
\code{\link{fwa_index_point}()},
\code{\link{fwa_locate_along}()},
\code{\link{fwa_locate_along_interval}()},
\code{\link{fwa_watershed_at_measure}()},
\code{\link{fwa_watershed_hex}()},
\code{\link{fwa_watershed_stream}()},
\code{\link{hydroshed}()}
}
\concept{functions}
