% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R, R/query-collection.R
\name{fwa_collection}
\alias{fwa_collection}
\alias{fwa_query_collection}
\title{Query Freshwater Atlas Collection}
\usage{
fwa_collection(
  collection_id,
  filter = NULL,
  limit = 10000,
  bbox = NULL,
  properties = NULL,
  transform = NULL,
  epsg = 4326
)

fwa_query_collection(
  collection_id,
  filter = NULL,
  limit = 10000,
  bbox = NULL,
  properties = NULL,
  transform = NULL,
  epsg = 4326
)
}
\arguments{
\item{collection_id}{A character string of the collection id.}

\item{filter}{A named vector or list of the filter(s) to apply, where the list
names correspond to column names and the list values correspond to the desired
value, e.g. \code{list(gnis_name = "Sangan River")}.}

\item{limit}{A positive whole number indicating the maximum number of features to return.}

\item{bbox}{A vector of four numbers indicating bounding box to limit output
features to, e.g. \code{c(minLon, minLat, maxLon, maxLat)}.}

\item{properties}{A vector of strings of the column names to include. If NULL (default), all columns are retained.}

\item{transform}{A character vector with the name of the \href{https://github.com/CrunchyData/pg_featureserv/blob/master/config/pg_featureserv.toml.example#L29}{\emph{valid transform}} function followed by the parameter values (e.g. c("ST_Simplify", 100))}

\item{epsg}{A positive whole number of the epsg to transform features to.}
}
\value{
An sf object.
}
\description{
Query a collection from the freshwater atlas.
}
\section{Functions}{
\itemize{
\item \code{fwa_collection}: Query Freshwater Atlas Collection

\lifecycle{soft-deprecated}
}}

\examples{
collection_id <- "whse_basemapping.fwa_stream_networks_sp"
filter <- list(gnis_name = "Sangan River")
fwa_query_collection(collection_id, filter = filter)
}
\seealso{
Other collections: 
\code{\link{fwa_collection_properties}()},
\code{\link{fwa_collections}()}
}
\concept{collections}
