% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-meta.R
\name{fwa_search_gnis_streams}
\alias{fwa_search_gnis_streams}
\title{FWA search named streams}
\usage{
fwa_search_gnis_streams(gnis_name = NULL, ignore_case = TRUE)
}
\arguments{
\item{gnis_name}{A regular expression to match gnis_name. If Null (default), returns all values.}

\item{ignore_case}{A flag indicating whether to ignore case in regular expression.}
}
\value{
A tibble
}
\description{
Match valid stream gnis_name with regular expression.
}
\examples{
fwa_search_gnis_streams("sangan")
}
