% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstream_trace.R
\name{fwa_upstream_trace}
\alias{fwa_upstream_trace}
\title{Freshwater Atlas Upstream Trace}
\usage{
fwa_upstream_trace(
  start_blue_line_key,
  start_measure = 0,
  tolerance = 1,
  bbox = NULL,
  properties = NULL,
  transform = NULL,
  epsg = 4326,
  nocache = getOption("fwa.nocache", FALSE)
)
}
\arguments{
\item{start_measure}{An integer of the distance in meters upstream from the river mouth to start from.}

\item{tolerance}{A number of the tolerance in m.}

\item{bbox}{A vector of four numbers indicating bounding box to limit output
features to, e.g. \code{c(minLon, minLat, maxLon, maxLat)}.}

\item{properties}{A vector of strings of the column names to include. If NULL (default), all columns are retained.}

\item{transform}{A character vector with the name of the \href{https://github.com/CrunchyData/pg_featureserv/blob/master/config/pg_featureserv.toml.example#L29}{\emph{valid transform}} function followed by the parameter values (e.g. c("ST_Simplify", 100))}

\item{epsg}{A positive whole number of the epsg to transform features to.}

\item{nocache}{A flag specifying whether or not to cache results.}
}
\value{
An sf object.
}
\description{
Returns the stream network upstream of the provided location
}
\examples{
fwa_upstream_trace(356308001, 10000)
}
\seealso{
\href{https://features.hillcrestgeo.ca/fwa/functions/fwa_upstreamtrace.html}{API documentation}

Other functions: 
\code{\link{fwa_index_point}()},
\code{\link{fwa_locate_along}()},
\code{\link{fwa_locate_along_interval}()},
\code{\link{fwa_network_trace}()},
\code{\link{fwa_network_trace_agg}()},
\code{\link{fwa_watershed_at_measure}()},
\code{\link{fwa_watershed_hex}()},
\code{\link{fwa_watershed_stream}()},
\code{\link{hydroshed}()}
}
\concept{functions}
