% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid-transforms.R
\name{fwa_valid_transforms}
\alias{fwa_valid_transforms}
\title{Freshwater Atlas Valid Transform Functions}
\usage{
fwa_valid_transforms()
}
\value{
A character vector of the valid transform functions.
}
\description{
Returns a character vector of the valid transform functions.
}
\details{
The valid transform functions are listed at \url{https://github.com/CrunchyData/pg_featureserv/blob/master/config/pg_featureserv.toml.example#L29}.
}
\examples{
fwa_valid_transforms()
}
