% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{fwa_watershed}
\alias{fwa_watershed}
\title{FWA watershed}
\usage{
fwa_watershed(blue_line_key, downstream_route_measure = 0, epsg = 4326)
}
\arguments{
\item{blue_line_key}{An integer of the stream blue line key.}

\item{downstream_route_measure}{An integer of the downstream route measure.}

\item{epsg}{An integer of the epsg to transform features to.}
}
\value{
A sf object
}
\description{
Get watershed upstream of a location on a stream.
}
\examples{
fwa_watershed(356308001)
}
