% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fwa_watershed_at_measure}
\alias{fwa_watershed_at_measure}
\title{FWA watershed at downstream route measure}
\usage{
fwa_watershed_at_measure(
  blue_line_key,
  downstream_route_measure = 0,
  epsg = 4326
)
}
\arguments{
\item{blue_line_key}{An integer of the stream blue line key.}

\item{downstream_route_measure}{An integer of the downstream route measure.}

\item{epsg}{An integer of the epsg to transform features to.}
}
\value{
A sf object
}
\description{
Provided a location as blue_line_key and downstream_route_measure,
return the entire watershed boundary upstream of the location.
}
\examples{
fwa_watershed_at_measure(356308001, downstream_route_measure = 10000)
}
