% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwapgr-package.R, R/pkgtemplate-package.R
\docType{package}
\name{fwapgr-package}
\alias{fwapgr-package}
\title{fwapgr: Client for BC Freshwater Atlas fwapg API}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Client for fwapg, a PostgreSQL database and web API for BC Freshwater Atlas data.

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Client for fwapg, a PostgreSQL database and web API for BC Freshwater Atlas data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/fwapgr}
  \item Report bugs at \url{https://github.com/poissonconsulting/fwapgr/issues}
}


Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/fwapgr}
  \item Report bugs at \url{https://github.com/poissonconsulting/fwapgr/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Dalgarno \email{seb@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-3658-4517}{ORCID})

Authors:
\itemize{
  \item Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})
}

Other contributors:
\itemize{
  \item Poisson Consulting [copyright holder, funder]
  \item Nadine Hussein \email{nadine@poissonconsulting.ca} (\href{https://orcid.org/0000-0003-4470-8361}{ORCID}) [contributor]
}

}
\keyword{internal}
