% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydroshed.R
\name{hydroshed}
\alias{hydroshed}
\title{Hydroshed}
\usage{
hydroshed(
  x,
  y,
  srid = 4326,
  bbox = NULL,
  properties = NULL,
  transform = NULL,
  epsg = 4326,
  nocache = getOption("fwa.nocache", TRUE)
)
}
\arguments{
\item{x}{A number of the x coordinate.}

\item{y}{A number of the y coordinate.}

\item{srid}{A positive whole number of the epsg of the coordinates.}

\item{bbox}{A vector of four numbers indicating bounding box to limit output
features to, e.g. \code{c(minLon, minLat, maxLon, maxLat)}.}

\item{properties}{A vector of strings of the column names to include. If NULL (default), all columns are retained.}

\item{transform}{A character vector with the name of the \href{https://github.com/CrunchyData/pg_featureserv/blob/master/config/pg_featureserv.toml.example#L29}{\emph{valid transform}} function followed by the parameter values (e.g. c("ST_Simplify", 100))}

\item{epsg}{A positive whole number of the epsg to transform features to.}

\item{nocache}{A flag specifying whether or not to cache results.}
}
\value{
An sf object.
}
\description{
Gets aggregated boundary of all \href{https://www.hydrosheds.org}{hydroshed}
polygons upstream of the provided location.
}
\examples{
hydroshed(x = -132.26, y = 53.36)
}
\seealso{
\href{https://features.hillcrestgeo.ca/fwa/functions/hydroshed.html}{API documentation}

Other functions: 
\code{\link{fwa_index_point}()},
\code{\link{fwa_locate_along_interval}()},
\code{\link{fwa_locate_along}()},
\code{\link{fwa_watershed_at_measure}()},
\code{\link{fwa_watershed_hex}()},
\code{\link{fwa_watershed_stream}()}
}
\concept{functions}
