% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgtemplate-package.R
\name{internal}
\alias{internal}
\title{Parameter Descriptions}
\arguments{
\item{collection_id}{A character string of the collection id.}

\item{filter}{A named vector or list of the filter(s) to apply, where the list
names correspond to column names and the list values correspond to the desired
value, e.g. \code{list(gnis_name = "Sangan River")}.}

\item{limit}{A positive whole number indicating the maximum number of features to return.}

\item{offset}{A positive whole number indicating the offset of start of returned results.}

\item{bbox}{A vector of four numbers indicating bounding box to limit output
features to, e.g. \code{c(minLon, minLat, maxLon, maxLat)}.}

\item{properties}{A vector of strings of the column names to include. If NULL (default), all columns are retained.}

\item{transform}{A character vector with the name of the transform function followed by the parameter values (e.g. c("ST_Simplify", 100))}

\item{epsg}{A positive whole number of the epsg to transform features to.}

\item{function_id}{A character string of the function id.}

\item{parameters}{A named list of the parameters and values for a specific function.}

\item{x}{A number of the x coordinate.}

\item{y}{A number of the y coordinate.}

\item{srid}{A positive whole number of the epsg of the coordinates.}

\item{tolerance}{A number of the tolerance in m.}

\item{blue_line_key}{A positive whole number of the stream blue line key.}

\item{downstream_route_measure}{A positive number of the river metre.}

\item{interval_length}{An integer of the interval distance in meters.}

\item{start_measure}{An integer of the distance in meters upstream from the river mouth to start from.}

\item{end_measure}{An integer of the distance in meters upstream from the river mouth to end at.}
}
\value{
An sf object.
}
\description{
Descriptions of the parameters.
}
\keyword{internal}
