% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{fwa_bbox}
\alias{fwa_bbox}
\title{BBox}
\usage{
fwa_bbox(
  table,
  filter = NULL,
  geom_column = "geom",
  epsg = 4326,
  schema = "whse_basemapping"
)
}
\arguments{
\item{table}{A character string of the table name.}

\item{filter}{A character string of the SQL WHERE statement to filter features by.}

\item{geom_column}{A character string of the name of the geometry column of the table.}

\item{epsg}{An integer of the epsg to transform features to.}

\item{schema}{A character string of the schema. Default 'whse_basemapping' is used for most layers.}
}
\value{
A sf object
}
\description{
Get bounding box of feature(s)
}
\examples{
fwa_bbox("fwa_stream_networks_sp", filter = "gnis_name = 'Cowichan River'")
}
