% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{fwa_collection}
\alias{fwa_collection}
\title{FWA feature}
\usage{
fwa_collection(
  table,
  filter = NULL,
  bbox = NULL,
  columns = NULL,
  epsg = 4326,
  limit = 10000,
  offset = 0,
  schema = "whse_basemapping"
)
}
\arguments{
\item{table}{A character string of the table name.}

\item{filter}{A named vector or list of the filter(s) to apply, where the list
names correspond to column names and the list values correspond to the desired
value, e.g. \code{list(gnis_name = "Sangan River")}.}

\item{bbox}{A vector of numbers indicating bounding box to limit output
features to, e.g. \code{c(minLon, minLat, maxLon, maxLat)}.}

\item{columns}{A vector of strings of the column names to include.}

\item{epsg}{A positive whole number of the epsg to transform features to.}

\item{limit}{A positive whole number indicating the maximum number of features to return.}

\item{offset}{A positive whole number indicating the offset of start of returned results.}

\item{schema}{A character string of the schema to prepend to collection_id.
Possible values include 'whse_basemapping', 'usgs', 'hydrosheds' and NULL. If NULL, nothing is prepended.}
}
\value{
A sf object
}
\description{
Get features from table.
}
\examples{
fwa_collection("fwa_stream_networks_sp", filter = list(gnis_name = 'Sangan River'))
}
