% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R, R/meta-collections.R
\name{fwa_collections}
\alias{fwa_collections}
\alias{fwa_meta_collections}
\title{Freshwater Atlas Collections}
\usage{
fwa_collections(nocache = getOption("fwa.nocache", TRUE))

fwa_meta_collections(nocache = getOption("fwa.nocache", TRUE))
}
\arguments{
\item{nocache}{A flag specifying whether or not to cache results.}
}
\value{
A tibble.
}
\description{
Get information about the collections.
}
\section{Functions}{
\itemize{
\item \code{fwa_meta_collections()}: Freshwater Atlas Collections

\lifecycle{soft-deprecated}

}}
\examples{
fwa_collections()
}
\seealso{
Other collections: 
\code{\link{fwa_collection_properties}()},
\code{\link{fwa_collection}()}
}
\concept{collections}
