% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fwa_function}
\alias{fwa_function}
\title{FWA function}
\usage{
fwa_function(
  function_id,
  parameters,
  limit = 10000,
  offset = 0,
  bbox = NULL,
  properties = NULL,
  transform = NULL,
  epsg = 4326
)
}
\arguments{
\item{function_id}{A character string of the function id}

\item{parameters}{A named list of the parameters and values for a specific function.}

\item{limit}{A positive whole number indicating the maximum number of features to return.}

\item{offset}{A positive whole number indicating the offset of start of returned results.}

\item{bbox}{A vector of numbers indicating bounding box to limit output
features to, e.g. \code{c(minLon, minLat, maxLon, maxLat)}.}

\item{properties}{A vector of strings of the column names to include. If NULL (default), all columns are retained.}

\item{transform}{A character string of the geometry transformation function pipeline to apply.}

\item{epsg}{A positive whole number of the epsg to transform features to.}
}
\value{
A sf object
}
\description{
General function API call
}
