% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{fwa_gnis_streams}
\alias{fwa_gnis_streams}
\title{FWA stream network}
\usage{
fwa_gnis_streams(
  gnis_name,
  columns = NULL,
  bounds = NULL,
  geom_column = "geom",
  epsg = 4326,
  named_streams = FALSE
)
}
\arguments{
\item{gnis_name}{A character string of the gnis_name.}

\item{columns}{A vector of strings of the column names to include.}

\item{bounds}{A vector of numbers indicating bounding box to limit output features to, e.g. \code{c(sw.lng, sw.lat, ne.lng, ne.lat)}.}

\item{geom_column}{A character string of the name of the geometry column of the table.}

\item{epsg}{An integer of the epsg to transform features to.}

\item{named_streams}{A flag indicating whether to get data from the 'fwa_named_streams' layer.
If FALSE, the 'fwa_stream_networks_sp' layer is used.}
}
\value{
A sf object
}
\description{
Get stream network features from gnis_name.
}
\examples{
\dontrun{
fwa_gnis_streams('Sangan River')
}
}
