% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index-point.R
\name{fwa_index_point}
\alias{fwa_index_point}
\title{Freshwater Atlas Index Point}
\usage{
fwa_index_point(
  x,
  y,
  srid = 4326,
  tolerance = 5000,
  limit = 1,
  bbox = NULL,
  properties = NULL,
  transform = NULL,
  epsg = 4326
)
}
\arguments{
\item{x}{A number of the x coordinate.}

\item{y}{A number of the y coordinate.}

\item{srid}{A positive whole number of the epsg of the coordinates.}

\item{tolerance}{A number of the tolerance in m.}

\item{limit}{A positive whole number indicating the maximum number of features to return.}

\item{bbox}{A vector of four numbers indicating bounding box to limit output
features to, e.g. \code{c(minLon, minLat, maxLon, maxLat)}.}

\item{properties}{A vector of strings of the column names to include. If NULL (default), all columns are retained.}

\item{transform}{A character vector with the name of the \href{https://github.com/CrunchyData/pg_featureserv/blob/master/config/pg_featureserv.toml.example#L29}{\emph{valid transform}} function followed by the parameter values (e.g. c("ST_Simplify", 100))}

\item{epsg}{A positive whole number of the epsg to transform features to.}
}
\value{
An sf object.
}
\description{
Gets the nearest point on a stream as the crow flies to the coordinates (within the tolerance).
}
\details{
To also return the second nearest point on a stream as the crow flies (within the tolerance)
set \code{limit = 2}.
}
\examples{
fwa_index_point(x = -132.26, y = 53.36)
}
\seealso{
\href{https://features.hillcrestgeo.ca/fwa/functions/fwa_indexpoint.html}{API documentation}

Other functions: 
\code{\link{fwa_locate_along_interval}()},
\code{\link{fwa_locate_along}()},
\code{\link{fwa_watershed_at_measure}()},
\code{\link{fwa_watershed_hex}()},
\code{\link{fwa_watershed_stream}()},
\code{\link{hydroshed}()}
}
\concept{functions}
