% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fwa_index_point}
\alias{fwa_index_point}
\title{FWA index point}
\usage{
fwa_index_point(
  x,
  y,
  srid,
  tolerance = 5000,
  num_features = 1,
  limit = 10000,
  offset = 0,
  bbox = NULL,
  properties = NULL,
  transform = NULL,
  epsg = srid
)
}
\arguments{
\item{x}{A number of the x coordinate.}

\item{y}{A number of the y coordinate.}

\item{srid}{An integer of the srid of the coordinate (e.g. epsg).}

\item{tolerance}{A number of the tolerance.}

\item{num_features}{An integer of the maximum number of features to return.}

\item{limit}{A positive whole number indicating the maximum number of features to return.}

\item{offset}{A positive whole number indicating the offset of start of returned results.}

\item{bbox}{A vector of numbers indicating bounding box to limit output
features to, e.g. \code{c(minLon, minLat, maxLon, maxLat)}.}

\item{properties}{A vector of strings of the column names to include. If NULL (default), all columns are retained.}

\item{transform}{A character string of the geometry transformation function pipeline to apply.}

\item{epsg}{A positive whole number of the epsg to transform features to.}
}
\value{
A sf object
}
\description{
Provided a point (x, y, srid), return the point along a stream(s) nearest to the point provided within specified tolerance.
}
\examples{
\dontrun{fwa_index_point(x = -132.26, y = 53.36, srid = 4326)}
}
\seealso{
\href{https://www.hillcrestgeo.ca/fwapg/functions/fwa_indexpoint.html}{fwa_indexpoint API documentation}
}
