% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{fwa_list_columns}
\alias{fwa_list_columns}
\title{FWA list columns}
\usage{
fwa_list_columns(table, schema = "whse_basemapping")
}
\arguments{
\item{schema}{A character string of the schema to prepend to collection_id.
Possible values include 'whse_basemapping', 'usgs', 'hydrosheds' and NULL. If NULL, nothing is prepended.}
}
\value{
A tibble
}
\description{
List columns in a table.
}
\examples{
fwa_list_columns("fwa_stream_networks_sp")
}
