% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-functions.R
\name{fwa_locate_along}
\alias{fwa_locate_along}
\title{Freshwater Atlas Locate Along}
\usage{
fwa_locate_along(
  blue_line_key,
  downstream_route_measure = 0,
  bbox = NULL,
  properties = NULL,
  transform = NULL,
  epsg = 4326
)
}
\arguments{
\item{blue_line_key}{A positive whole number of the stream blue line key.}

\item{downstream_route_measure}{A positive number of the river metre.}

\item{bbox}{A vector of four numbers indicating bounding box to limit output
features to, e.g. \code{c(minLon, minLat, maxLon, maxLat)}.}

\item{properties}{A vector of strings of the column names to include. If NULL (default), all columns are retained.}

\item{transform}{A character vector with the name of the transform function followed by the parameter values (e.g. c("ST_Simplify", 100))}

\item{epsg}{A positive whole number of the epsg to transform features to.}
}
\value{
An sf object.
}
\description{
Gets the point on a stream corresponding to the river metre
(downstream route measure).
}
\examples{
fwa_locate_along(356308001, downstream_route_measure = 10000)
}
\seealso{
\href{https://www.hillcrestgeo.ca/fwapg/functions/fwa_locatealong.html}{API documentation}

Other functions: 
\code{\link{fwa_index_point}()},
\code{\link{fwa_locate_along_interval}()},
\code{\link{fwa_watershed_at_measure}()},
\code{\link{fwa_watershed_hex}()},
\code{\link{fwa_watershed_stream}()}
}
\concept{functions}
