% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-collections.R
\name{fwa_meta_collections}
\alias{fwa_meta_collections}
\title{FWA Collections Metadata}
\usage{
fwa_meta_collections()
}
\value{
A tibble.
}
\description{
Get information about the collections.
}
\details{
Deprecated for \code{fwa_collections()}.
}
\examples{
fwa_meta_collections()
}
\seealso{
Other collections: 
\code{\link{fwa_collections_properties}()},
\code{\link{fwa_collections}()},
\code{\link{fwa_collection}()},
\code{\link{fwa_meta_properties}()}
}
\concept{collections}
