% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{fwa_meta_collections}
\alias{fwa_meta_collections}
\title{FWA Collections Metadata}
\usage{
fwa_meta_collections()
}
\value{
A tibble.
}
\description{
Get information about the collections.
}
\examples{
fwa_meta_collections()
}
\seealso{
Other metadata: 
\code{\link{fwa_meta_properties}()}
}
\concept{metadata}
