% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-properties.R
\name{fwa_meta_properties}
\alias{fwa_meta_properties}
\title{FWA Property Metadata}
\usage{
fwa_meta_properties(collection_id)
}
\arguments{
\item{collection_id}{A character string of the collection id.}
}
\value{
A tibble.
}
\description{
Get information about the column names of a collection.
}
\details{
Soft_deprecated for \code{\link[=fwa_collections_properties]{fwa_collections_properties()}}
}
\examples{
fwa_meta_properties("whse_basemapping.fwa_stream_networks_sp")
}
\seealso{
Other metadata: 
\code{\link{fwa_collections_properties}()},
\code{\link{fwa_collections}()},
\code{\link{fwa_meta_collections}()}
}
\concept{metadata}
