% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{fwa_meta_properties}
\alias{fwa_meta_properties}
\title{FWA property metadata}
\usage{
fwa_meta_properties(collection_id)
}
\arguments{
\item{collection_id}{A character string of the collection id.}
}
\value{
A tibble
}
\description{
Provide column names of collection.
}
\examples{
\dontrun{fwa_meta_properties("whse_basemapping.fwa_stream_networks_sp")}
}
\seealso{
Other metadata: 
\code{\link{fwa_meta_collections}()}
}
\concept{metadata}
