% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{fwa_watershed}
\alias{fwa_watershed}
\title{FWA watershed}
\usage{
fwa_watershed(blue_line_key, downstream_route_measure = 0, srid = 4326)
}
\arguments{
\item{blue_line_key}{An integer of the blue line key of the stream for which to generate the watershed.}

\item{downstream_route_measure}{An integer of the downstream route measure at the watershed terminus / pour point.}

\item{srid}{An integer of the SRID for the returned watershed. The default is 4326 WGS84 Lat/Lng.}
}
\value{
A sf object
}
\description{
Get watershed upstream of a location on a stream.
}
\examples{
fwa_watershed(356308001)
}
