% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed-at-measure.R
\name{fwa_watershed_at_measure}
\alias{fwa_watershed_at_measure}
\title{Freshwater Atlas Watershed at Downstream Route Measure}
\usage{
fwa_watershed_at_measure(
  blue_line_key,
  downstream_route_measure = 0,
  bbox = NULL,
  properties = NULL,
  transform = NULL,
  epsg = 4326
)
}
\arguments{
\item{blue_line_key}{A positive whole number of the stream blue line key.}

\item{downstream_route_measure}{A positive number of the river metre.}

\item{bbox}{A vector of four numbers indicating bounding box to limit output
features to, e.g. \code{c(minLon, minLat, maxLon, maxLat)}.}

\item{properties}{A vector of strings of the column names to include. If NULL (default), all columns are retained.}

\item{transform}{A character vector with the name of the \href{https://github.com/CrunchyData/pg_featureserv/blob/master/config/pg_featureserv.toml.example#L29}{\emph{valid transform}} function followed by the parameter values (e.g. c("ST_Simplify", 100))}

\item{epsg}{A positive whole number of the epsg to transform features to.}
}
\value{
An sf object.
}
\description{
Gets a polygon of the watershed at a specified river metre.
}
\details{
The function selects and aggregate all the fundamental watershed polygons
upstream of the fundamental watershed in which the point lies.

Depending on where the point is in the fundamental watershed one of four
things can happen to the fundamental watershed.
\enumerate{
\item If the point is < 50 m upstream from the bottom of the fundamental
watershed then the fundamental watershed is included in the aggregation
(recorded as KEEP in the refine_method property).
\item Else if the point is < 100m downstream from the top of the fundamental
watershed then the fundamental watershed is not included in the aggregation
(recorded as DROP in the refine_method property).
\item Else if the point is on a polygonal river/canal the fundamental watershed
is cut across the banks of the river before being included in the aggregation
(recorded as CUT in the refine_method property).
\item Otherwise something else happens (recorded as DEM in the refine_method
property).
}
}
\examples{
fwa_watershed_at_measure(356308001, downstream_route_measure = 10000)
}
\seealso{
\href{https://features.hillcrestgeo.ca/fwa/functions/fwa_watershedatmeasure.html}{API documentation}

Other functions: 
\code{\link{fwa_index_point}()},
\code{\link{fwa_locate_along_interval}()},
\code{\link{fwa_locate_along}()},
\code{\link{fwa_watershed_hex}()},
\code{\link{fwa_watershed_stream}()},
\code{\link{hydroshed}()}
}
\concept{functions}
