% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{fwa_watershed_hex}
\alias{fwa_watershed_hex}
\title{FWA watershed hex}
\usage{
fwa_watershed_hex(blue_line_key, downstream_route_measure = 0, srid = 4326)
}
\arguments{
\item{blue_line_key}{An integer of the stream blue line key.}

\item{downstream_route_measure}{An integer of the downstream route measure.}

\item{srid}{An integer of the SRID for the returned feature. The default is 4326 WGS84 Lat/Lng.}
}
\value{
A sf object
}
\description{
Get hex grid covering first order watershed.
}
\examples{
fwa_watershed_hex(356308001)
}
