% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fwa_watershed_hex}
\alias{fwa_watershed_hex}
\title{FWA watershed hex}
\usage{
fwa_watershed_hex(
  blue_line_key,
  downstream_route_measure = 0,
  limit = 10000,
  offset = 0,
  bbox = NULL,
  properties = NULL,
  transform = NULL,
  epsg = 4326
)
}
\arguments{
\item{blue_line_key}{An integer of the stream blue line key.}

\item{downstream_route_measure}{An integer of the downstream route measure.}

\item{limit}{A positive whole number indicating the maximum number of features to return.}

\item{offset}{A positive whole number indicating the offset of start of returned results.}

\item{bbox}{A vector of numbers indicating bounding box to limit output
features to, e.g. \code{c(minLon, minLat, maxLon, maxLat)}.}

\item{properties}{A vector of strings of the column names to include. If NULL (default), all columns are retained.}

\item{transform}{A character string of the geometry transformation function pipeline to apply.}

\item{epsg}{A positive whole number of the epsg to transform features to.}
}
\value{
A sf object
}
\description{
Provided a location as blue_line_key and downstream_route_measure, return a 25m hexagon grid covering first order watershed in which location lies.
}
\examples{
\dontrun{fwa_watershed_hex(356308001, downstream_route_measure = 10000)}
}
\seealso{
\href{https://www.hillcrestgeo.ca/fwapg/functions/fwa_watershedhex.html}{fwa_watershedhex API documentation}
}
