% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed-hex.R
\name{fwa_watershed_hex}
\alias{fwa_watershed_hex}
\title{Freshwater Atlas Watershed Hex}
\usage{
fwa_watershed_hex(
  blue_line_key,
  downstream_route_measure = 0,
  limit = 10000,
  bbox = NULL,
  properties = NULL,
  transform = NULL,
  epsg = 4326,
  nocache = getOption("fwa.nocache", FALSE)
)
}
\arguments{
\item{blue_line_key}{A positive whole number of the stream blue line key.}

\item{downstream_route_measure}{A positive number of the river metre.}

\item{limit}{A positive whole number indicating the maximum number of features to return.}

\item{bbox}{A vector of four numbers indicating bounding box to limit output
features to, e.g. \code{c(minLon, minLat, maxLon, maxLat)}.}

\item{properties}{A vector of strings of the column names to include. If NULL (default), all columns are retained.}

\item{transform}{A character vector with the name of the \href{https://github.com/CrunchyData/pg_featureserv/blob/master/config/pg_featureserv.toml.example#L29}{\emph{valid transform}} function followed by the parameter values (e.g. c("ST_Simplify", 100))}

\item{epsg}{A positive whole number of the epsg to transform features to.}

\item{nocache}{A flag specifying whether or not to cache results.}
}
\value{
An sf object.
}
\description{
Gets a 25m hexagon grid covering the fundamental watershed in which the river metre lies.
}
\examples{
\dontrun{
fwa_watershed_hex(356308001, downstream_route_measure = 10000)
}
}
\seealso{
\href{https://features.hillcrestgeo.ca/fwa/functions/fwa_watershedhex.html}{API documentation}

Other functions: 
\code{\link{fwa_index_point}()},
\code{\link{fwa_locate_along_interval}()},
\code{\link{fwa_locate_along}()},
\code{\link{fwa_watershed_at_measure}()},
\code{\link{fwa_watershed_stream}()},
\code{\link{hydroshed}()}
}
\concept{functions}
