% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed-stream.R
\name{fwa_watershed_stream}
\alias{fwa_watershed_stream}
\title{Freshwater Atlas Watershed Stream}
\usage{
fwa_watershed_stream(
  blue_line_key,
  downstream_route_measure = 0,
  bbox = NULL,
  properties = NULL,
  transform = NULL,
  epsg = 4326,
  nocache = getOption("fwa.nocache", TRUE)
)
}
\arguments{
\item{blue_line_key}{A positive whole number of the stream blue line key.}

\item{downstream_route_measure}{A positive number of the river metre.}

\item{bbox}{A vector of four numbers indicating bounding box to limit output
features to, e.g. \code{c(minLon, minLat, maxLon, maxLat)}.}

\item{properties}{A vector of strings of the column names to include. If NULL (default), all columns are retained.}

\item{transform}{A character vector with the name of the \href{https://github.com/CrunchyData/pg_featureserv/blob/master/config/pg_featureserv.toml.example#L29}{\emph{valid transform}} function followed by the parameter values (e.g. c("ST_Simplify", 100))}

\item{epsg}{A positive whole number of the epsg to transform features to.}

\item{nocache}{A flag specifying whether or not to cache results.}
}
\value{
An sf object.
}
\description{
Gets the upstream stream segment within the fundamental watershed.
}
\details{
This is useful for adding stream network that was not included in the
watershed at measure.
}
\examples{
fwa_watershed_stream(356308001, downstream_route_measure = 10000)
}
\seealso{
\href{https://features.hillcrestgeo.ca/fwa/functions/fwa_watershedstream.html}{API documentation}

Other functions: 
\code{\link{fwa_index_point}()},
\code{\link{fwa_locate_along_interval}()},
\code{\link{fwa_locate_along}()},
\code{\link{fwa_watershed_at_measure}()},
\code{\link{fwa_watershed_hex}()},
\code{\link{hydroshed}()}
}
\concept{functions}
