% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fwa_watershed_stream}
\alias{fwa_watershed_stream}
\title{FWA watershed stream}
\usage{
fwa_watershed_stream(
  blue_line_key,
  downstream_route_measure = 0,
  limit = 100,
  offset = 0,
  bbox = NULL,
  properties = NULL,
  transform = NULL,
  epsg = 4326
)
}
\arguments{
\item{blue_line_key}{An integer of the stream blue line key.}

\item{downstream_route_measure}{An integer of the downstream route measure.}

\item{limit}{A positive whole number indicating the maximum number of features to return.}

\item{offset}{A positive whole number indicating the offset of start of returned results.}

\item{bbox}{A vector of numbers indicating bounding box to limit output
features to, e.g. \code{c(minLon, minLat, maxLon, maxLat)}.}

\item{properties}{A vector of strings of the column names to include. If NULL (default), all columns are retained.}

\item{transform}{A vector of strings of the geometry transform function followed by parameter values (e.g. c("ST_Simplify", 100))}

\item{epsg}{A positive whole number of the epsg to transform features to.}
}
\value{
A sf object
}
\description{
Provided a location as blue_line_key and downstream_route_measure, return stream segments upstream, within the same first order watershed.
}
\examples{
\dontrun{fwa_watershed_stream(356308001, downstream_route_measure = 10000)}
}
\seealso{
\href{https://www.hillcrestgeo.ca/fwapg/functions/fwa_watershedstream.html}{fwa_watershedstream API documentation}

Other functions: 
\code{\link{fwa_index_point}()},
\code{\link{fwa_locate_along_interval}()},
\code{\link{fwa_locate_along}()},
\code{\link{fwa_watershed_at_measure}()},
\code{\link{fwa_watershed_hex}()}
}
\concept{functions}
