% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgtemplate-package.R
\docType{package}
\name{fwapgr-package}
\alias{fwapgr}
\alias{fwapgr-package}
\title{fwapgr: FWA PostgreSQL database API wrappers}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

FWA PostgreSQL database API wrappers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/fwapgr}
  \item Report bugs at \url{https://github.com/poissonconsulting/fwapgr/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Dalgarno \email{seb@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-3658-4517}{ORCID})

Other contributors:
\itemize{
  \item Poisson Consulting [copyright holder, funder]
}

}
\keyword{internal}
