% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgtemplate-package.R
\name{internal}
\alias{internal}
\title{Parameter Descriptions}
\arguments{
\item{bbox}{A vector of four numbers indicating bounding box to limit output
features to, e.g. \code{c(minLon, minLat, maxLon, maxLat)}.}

\item{blue_line_key}{A positive whole number of the stream blue line key.}

\item{collection_id}{A character string of the collection id.}

\item{downstream_route_measure}{A positive number of the river metre.}

\item{end_measure}{An integer of the distance in meters upstream from the river mouth to end at.}

\item{epsg}{A positive whole number of the epsg to transform features to.}

\item{function_id}{A character string of the function id.}

\item{filter}{A named vector or list of the filter(s) to apply, where the list
names correspond to column names and the list values correspond to the desired
value, e.g. \code{list(gnis_name = "Sangan River")}.}

\item{interval_length}{An integer of the interval distance in meters.}

\item{limit}{A positive whole number indicating the maximum number of features to return.}

\item{offset}{A positive whole number indicating the offset of start of returned results.}

\item{parameters}{A named list of the parameters and values for a specific function.}

\item{properties}{A vector of strings of the column names to include. If NULL (default), all columns are retained.}

\item{sortby}{A string that sorts the response items by a property.
Default is ascending but property name can be prepended with
'-' to indicate descending.}

\item{groupby}{A string of the property name (i.e., column name) to group
response features by. This would typically be used with an aggregate
transform function.}

\item{srid}{A positive whole number of the epsg of the coordinates.}

\item{start_measure}{An integer of the distance in meters upstream from the river mouth to start from.}

\item{tolerance}{A number of the tolerance in m.}

\item{transform}{A character vector with the name of the \href{https://github.com/CrunchyData/pg_featureserv/blob/master/config/pg_featureserv.toml.example#L29}{\emph{valid transform}} function followed by the parameter values (e.g. c("ST_Simplify", 100))}

\item{x}{A number of the x coordinate.}

\item{y}{A number of the y coordinate.}

\item{nocache}{A flag specifying whether or not to cache results.}
}
\value{
An sf object.
}
\description{
Descriptions of the parameters and a typical return value for the provided functions.
}
\keyword{internal}
